tests/submission/Google/NavigationTiming/navigation_timing_attributes.htm

<!doctype html>
<html class="a">
 <head>
  <title>window.performance.timing attributes</title>
  <script src="../../../resources/testharness.js"></script>
</head>
<body class="a">
  <div id="log"></div>
  <script>

  var nt_namespace = window.performance
  function test_window_performance() {
    assert_true(nt_namespace != undefined, "window.performance does not exist.");
  }
  test(test_window_performance, "window.performance is defined.");

  function test_timing() {
    assert_true(nt_namespace.timing != undefined,
                "window.performance.timing is defined.");
  }
  test(test_timing, "window.performance.timing is defined.");

  // All the attributes under window.performance.timing exist. 
  var timing = nt_namespace.timing;
  var timing_attribute_list = new Array();
  timing_attribute_list[0] = "navigationStart";
  timing_attribute_list[1] = "unloadEventStart";
  timing_attribute_list[2] = "unloadEventEnd";
  timing_attribute_list[3] = "redirectStart";
  timing_attribute_list[4] = "redirectEnd";
  timing_attribute_list[5] = "fetchStart";
  timing_attribute_list[6] = "domainLookupStart";
  timing_attribute_list[7] = "domainLookupEnd";
  timing_attribute_list[8] = "connectStart";
  timing_attribute_list[9] = "connectEnd";
  timing_attribute_list[10] = "handshakeStart";
  timing_attribute_list[11] = "requestStart";
  timing_attribute_list[12] = "requestEnd";
  timing_attribute_list[13] = "responseEnd";
  timing_attribute_list[14] = "domLoading";
  timing_attribute_list[15] = "domInteractive";
  timing_attribute_list[16] = "domContentLoadedEventStart";
  timing_attribute_list[17] = "domContentLoadedEventEnd";
  timing_attribute_list[18] = "domComplete";
  timing_attribute_list[19] = "loadEventStart";
  timing_attribute_list[20] = "loadEventEnd";
  var a_counter = 0;
  function test_timing_attributes() {
    assert_true(timing[timing_attribute_list[a_counter]] != undefined,
                timing_attribute_list[a_counter] + " should be defined.");
  }
  for (a_counter = 0; a_counter < timing_attribute_list.length; a_counter++) {
    test(test_timing_attributes,
         timing_attribute_list[a_counter] + " should be defined.");
  }

  function test_navigation() {
    assert_true(nt_namespace.navigation != undefined,
                "window.performance.navigation is defined.");
  }
  test(test_navigation, "window.performance.navigation is defined.");

  // All the attributes under window.performance.navigation exist.
  var navigation = nt_namespace.navigation;
  var nav_attribute_list = new Array();
  nav_attribute_list[0] = "type";
  nav_attribute_list[1] = "redirectCount";

  a_counter = 0;
  function test_navigation_attributes() {
    assert_true(navigation[nav_attribute_list[a_counter]] != undefined,
                nav_attribute_list[a_counter] + " should be defined.");
  }
  for (a_counter = 0; a_counter < nav_attribute_list.length; a_counter++) {
    test(test_navigation_attributes,
         nav_attribute_list[a_counter] + " should be defined.");
  }

  </script>
</body>
</html>
